Adjust `prettier` pre-commit hook version
Pin the `prettier` hook on the last v2 tag available (`v2.7.1`) and add
a hack to get the most recent v2 release available for `prettier` since
the hook is a pre-commit mirror hook. The v3 release of `prettier` adds
logic for formatting JavaScript/TypeScript files that causes build
failures due to the age of this project. These changes are not present
in v2 releases of `prettier`. Lastly run the prettier hook against the
project to update any files appropriately.