Restore an intentionally minified JSON
This restores a JSON that was pretty formatted to its minified state
and updated the pre-commit configuration to maintain that state. Since
disabling JSON pretty formatting is not something we typically do I
decided to make the exclusion only apply to this specific file and
added a comment directing any future exclusions to be explicit as well.

Co-authored-by: Shane Frasier <maverick@maverickdolphin.com>