Add a npm shrinkwrap file to the configuration
The build for this project fails now due to a change in the 2.53.48
version of the @types/selenium-webdriver package. This package is not
an explicit dependency of this project, but is instead part of the
dependency tree for the protractor package which _is_ a project
dependency. Since this project is so old, and such future breakages
have a reasonable likelihood of recurring, we will begin to shrinkwrap
the project's dependencies.

Due to being unable to successfully perform a `npm install` with the
current configuration I had to manually create the npm-shrinkwrap.json
file. The only pin included in this initial version is for the
problematic package. Once an image can be created and tested the
npm-shrinkwrap.json file should be re-generated to capture all version
pins that would be included by running `npm shrinkwrap`.