Add a bower configuration file
The default bower registry URL `registry.bower.io` throws an expired
certificate error in our Docker configuration. We get around this by
directing bower to use the legacy registry URL: `bower.herokuapp.com`.
If/when this change fails then the only option we'll have with the
existing configuration is to disable the SSL check for bower
completely. This should be done by changing the `registry` key to the
default and adding a `"strict-ssl": false` item.